requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
pydantic==2.5.2
pydantic-settings==2.1.0
httpx>=0.24.0,<0.25.0
python-dotenv==1.0.0
pillow==10.1.0
openai==1.3.5
requests==2.31.0
email-validator==2.1.0.post1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
psutil==5.9.6
boto3==1.34.7
aiofiles==23.2.1
runwayml
mailjet-rest==1.3.4
asyncpg==0.29.0
jinja2==3.1.2
stripe==7.10.0
pytz==2024.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0 